@@ -9,7 +9,7 @@
     "@heroku-cli/color": "2.0.1",
     "@heroku-cli/command": "^11.3.1",
     "@heroku-cli/notifications": "^1.2.4",
-    "@heroku-cli/plugin-ps-exec": "^2.4.0",
+    "@heroku-cli/plugin-ps-exec": "2.6.0-beta.0",
     "@heroku-cli/schema": "^1.0.25",
     "@heroku/buildpack-registry": "^1.0.1",
     "@heroku/eventsource": "^1.0.7",
@@ -76,7 +76,6 @@
     "tslib": "1.14.1",
     "tunnel-ssh": "4.1.6",
     "urijs": "^1.19.11",
-    "valid-url": "^1.0.9",
     "validator": "^13.7.0",
     "word-wrap": "^1.2.5",
     "ws": "^6.2.2"
@@ -94,7 +93,7 @@
     "@types/inquirer": "^8.2.10",
     "@types/lodash": "^4.14.123",
     "@types/mocha": "^10.0.6",
-    "@types/node": "16.11.7",
+    "@types/node": "20.14.8",
     "@types/node-fetch": "^2.1.6",
     "@types/phoenix": "^1.4.0",
     "@types/proxyquire": "^1.3.28",
@@ -121,7 +120,7 @@
     "mocha": "^9.2.2",
     "nock": "^13.5.1",
     "nyc": "^15.1.0",
-    "oclif": "4.4.7",
+    "oclif": "4.14.36",
     "proxyquire": "^2.1.0",
     "qqjs": "0.3.11",
     "read-pkg": "^4.0.1",
@@ -134,7 +133,7 @@
     "typescript": "4.8.4"
   },
   "engines": {
-    "node": "~16.20.0"
+    "node": "~20.x"
   },
   "files": [
     "/autocomplete-scripts",
@@ -340,7 +339,7 @@
     },
     "update": {
       "node": {
-        "version": "16.20.2"
+        "version": "20.17.0"
       },
       "s3": {
         "xz": true,
