@@ -38,7 +38,7 @@ export default class Set extends Command {
     if (flags.index === undefined) {
       spliceIndex = 0
     } else {
-      // eslint-disable-next-line unicorn/no-array-callback-reference, unicorn/no-array-method-this-argument
+      // eslint-disable-next-line unicorn/no-array-method-this-argument
       const foundIndex = buildpackCommand.findIndex(buildpacks, flags.index)
       spliceIndex = (foundIndex === -1) ? buildpacks.length : foundIndex
     }
