@@ -1,4 +1,4 @@
-import {Args, ux} from '@oclif/core'
+import {ux} from '@oclif/core'
 import color from '@heroku-cli/color'
 import * as Heroku from '@heroku-cli/schema'
 import {flags, Command} from '@heroku-cli/command'
