@@ -5,7 +5,7 @@ import * as Heroku from '@heroku-cli/schema'
 import pghost from '../../lib/pg/host'
 import {getAddon, all} from '../../lib/pg/fetcher'
 import {configVarNamesFromValue, databaseNameFromUrl} from '../../lib/pg/util'
-import {AddOnAttachmentWithConfigVarsAndPlan, AddOnWithRelatedData, PgDatabaseTenant, TenantInfo} from '../../lib/pg/types'
+import {AddOnAttachmentWithConfigVarsAndPlan, AddOnWithRelatedData, PgDatabaseTenant} from '../../lib/pg/types'
 import {nls} from '../../nls'
 
 type DBObject = {
