@@ -2,44 +2,74 @@ import {Command, flags} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
 import color from '@heroku-cli/color'
 import {Args, ux} from '@oclif/core'
-
+import {ProcessTypeCompletion} from '@heroku-cli/command/lib/completions'
+import heredoc from 'tsheredoc'
 export default class Restart extends Command {
-  static description = 'restart app dynos'
+  static description = heredoc(`
+    restart an app dyno or process type
+    if neither --dyno nor --type are specified, restarts all dynos on app
+  `)
+
   static topic = 'ps'
   static aliases = ['dyno:restart']
   static hiddenAliases = ['restart']
 
   static examples = [
-    '$ heroku ps:restart web.1',
-    '$ heroku ps:restart web',
-    '$ heroku ps:restart',
+    '$ heroku ps:restart --app myapp --dyno-name web.1',
+    '$ heroku ps:restart --app myapp --process-type web',
+    '$ heroku ps:restart --app myapp',
   ]
 
-  static help = 'if DYNO is not specified, restarts all dynos on app'
-
   static args = {
-    dyno: Args.string({required: false, description: 'ID of the dyno'}),
+    dyno: Args.string({description: 'name of the dyno to restart', required: false, deprecated: true}),
   }
 
   static flags = {
     app: flags.app({required: true}),
     remote: flags.remote(),
+    'dyno-name': flags.string({
+      char: 'd',
+      description: 'name of the dyno to restart',
+    }),
+    'process-type': flags.string({
+      char: 'p',
+      description: 'name of the process type to restart',
+      completion: ProcessTypeCompletion,
+      exclusive: ['dyno-name'],
+    }),
   }
 
   async run() {
     const {args, flags} = await this.parse(Restart)
-
     const app = flags.app
-    const dyno = args.dyno
-
+    const dyno = flags['dyno-name'] || args.dyno
+    const type = flags['process-type']
     let msg = 'Restarting'
+    let restartUrl
 
-    if (dyno) msg += ` ${color.cyan(dyno)}`
-    msg += (dyno && dyno.includes('.')) ? ' dyno' : ' dynos'
+    if (type) {
+      msg += ` all ${color.cyan(type)} dynos`
+      restartUrl = `/apps/${app}/formations/${encodeURIComponent(type)}`
+    } else if (dyno) {
+      if (args.dyno) {
+        ux.warn(`DYNO is a deprecated argument. Use ${color.cmd('--dyno-name')} or ${color.cmd('--process-type')} instead.`)
+      }
+
+      msg += ` dyno ${color.cyan(dyno)}`
+      restartUrl = `/apps/${app}/dynos/${encodeURIComponent(dyno)}`
+    } else {
+      msg += ' all dynos'
+      restartUrl = `/apps/${app}/dynos`
+    }
+
     msg += ` on ${color.app(app)}`
 
     ux.action.start(msg)
-    await this.heroku.delete<Heroku.Dyno>(dyno ? `/apps/${app}/dynos/${encodeURIComponent(dyno)}` : `/apps/${app}/dynos`)
+    await this.heroku.delete<Heroku.Dyno>(restartUrl, {
+      headers: {
+        Accept: 'application/vnd.heroku+json; version=3.sdk',
+      },
+    })
     ux.action.stop()
   }
 }
