@@ -4,7 +4,10 @@ import * as Heroku from '@heroku-cli/schema'
 export const findRelease = async function (heroku: APIClient, app: string, search: (releases: Heroku.Release[]) => Heroku.Release) {
   const {body: releases} = await heroku.request<Heroku.Release[]>(`/apps/${app}/releases`, {
     partial: true,
-    headers: {Range: 'version ..; max=10, order=desc'},
+    headers: {
+      Range: 'version ..; max=10, order=desc',
+      Accept: 'application/vnd.heroku+json; version=3.sdk',
+    },
   })
 
   return search(releases)
@@ -14,7 +17,11 @@ export const getRelease = async function (heroku: APIClient, app: string, releas
   let id = release.toLowerCase()
   id = id.startsWith('v') ? id.slice(1) : id
 
-  const {body: releaseResponse} = await heroku.get<Heroku.Release>(`/apps/${app}/releases/${id}`)
+  const {body: releaseResponse} = await heroku.get<Heroku.Release>(`/apps/${app}/releases/${id}`, {
+    headers: {
+      Accept: 'application/vnd.heroku+json; version=3.sdk',
+    },
+  })
 
   return releaseResponse
 }
@@ -29,7 +36,7 @@ export const findByLatestOrId = async function (heroku: APIClient, app: string, 
 
 export const findByPreviousOrId = async function (heroku: APIClient, app: string, release = 'previous') {
   if (release === 'previous') {
-    return findRelease(heroku, app, releases => releases.filter(r => r.status === 'succeeded')[1])
+    return findRelease(heroku, app, releases => releases.filter(r => r.eligible_for_rollback)[1])
   }
 
   return getRelease(heroku, app, release)
