@@ -1,37 +1,23 @@
 import {APIClient} from '@heroku-cli/command'
 import {ux} from '@oclif/core'
-import HTTP from 'http-call'
-import {URL} from 'url'
-
+import color from '@heroku-cli/color'
 import colorize from './colorize'
-import liner from './line-transform'
+import {isFirApp} from '../apps/generation'
+import {LogSession} from '../types/fir'
 
 const EventSource = require('@heroku/eventsource')
 
 interface LogDisplayerOptions {
-  app: string;
-  dyno?: string;
-  lines?: number;
-  tail: boolean;
-  source?: string;
+  app: string,
+  dyno?: string
+  lines?: number
+  source?: string
+  tail: boolean
+  type?: string
 }
 
-async function readLogsV1(logplexURL: string) {
-  const {response} = await HTTP.stream(logplexURL)
-  return new Promise(function (resolve, reject) {
-    response.setEncoding('utf8')
-    liner.setEncoding('utf8')
-    response.pipe(liner)
-    liner.on('data', line => ux.log(colorize(line)))
-    response.on('end', resolve)
-    response.on('error', reject)
-  })
-}
-
-function readLogsV2(logplexURL: string) {
+function readLogs(logplexURL: string, isTail: boolean, recreateSessionTimeout?: number) {
   return new Promise<void>(function (resolve, reject) {
-    const u = new URL(logplexURL)
-    const isTail = u.searchParams.get('tail') === 'true'
     const userAgent = process.env.HEROKU_DEBUG_USER_AGENT || 'heroku-run'
     const proxy = process.env.https_proxy || process.env.HTTPS_PROXY
     const es = new EventSource(logplexURL, {
@@ -41,12 +27,12 @@ function readLogsV2(logplexURL: string) {
       },
     })
 
-    es.addEventListener('error', function (err: { status: number; message: any }) {
+    es.addEventListener('error', function (err: { status?: number; message?: string | null }) {
       if (err && (err.status || err.message)) {
         const msg = (isTail && (err.status === 404 || err.status === 403)) ?
           'Log stream timed out. Please try again.' :
-          `Logs eventsource failed with: ${err.status} ${err.message}`
-        reject(msg)
+          `Logs eventsource failed with: ${err.status}${err.message ? ` ${err.message}` : ''}`
+        reject(new Error(msg))
         es.close()
       }
 
@@ -63,17 +49,14 @@ function readLogsV2(logplexURL: string) {
         ux.log(colorize(line))
       })
     })
+
+    if (isTail && recreateSessionTimeout) {
+      setTimeout(() => {
+        reject(new Error('Fir log stream timeout'))
+        es.close()
+      }, recreateSessionTimeout)
+    }
   })
-}
-
-function readLogs(logplexURL: string) {
-  const u = new URL(logplexURL)
-
-  if (u.searchParams.has('srv')) {
-    return readLogsV1(logplexURL)
-  }
-
-  return readLogsV2(logplexURL)
 }
 
 async function logDisplayer(heroku: APIClient, options: LogDisplayerOptions) {
@@ -85,16 +68,48 @@ async function logDisplayer(heroku: APIClient, options: LogDisplayerOptions) {
     }
   })
 
-  const response: { body: { logplex_url: string}} = await heroku.post(`/apps/${options.app}/log-sessions`, {
-    body: {
-      tail: options.tail,
-      dyno: options.dyno,
-      source: options.source,
-      lines: options.lines,
-    },
-  })
+  const firApp = await isFirApp(options.app, heroku)
+  const isTail = firApp || options.tail
 
-  return readLogs(response.body.logplex_url)
+  const requestBodyParameters = {
+    source: options.source,
+  }
+
+  if (firApp) {
+    process.stderr.write(color.cyan.bold('Fetching logs...\n\n'))
+    Object.assign(requestBodyParameters, {
+      dyno: options.dyno,
+      type: options.type,
+    })
+  } else {
+    Object.assign(requestBodyParameters, {
+      dyno: options.dyno || options.type,
+      lines: options.lines,
+      tail: options.tail,
+    })
+  }
+
+  let recreateLogSession = false
+  do {
+    const {body: logSession} = await heroku.post<LogSession>(`/apps/${options.app}/log-sessions`, {
+      body: requestBodyParameters,
+      headers: {Accept: 'application/vnd.heroku+json; version=3.sdk'},
+    })
+
+    try {
+      await readLogs(
+        logSession.logplex_url,
+        isTail,
+        firApp ? Number(process.env.HEROKU_LOG_STREAM_TIMEOUT || '15') * 60 * 1000 : undefined,
+      )
+    } catch (error: unknown) {
+      const {message} = error as Error
+      if (message === 'Fir log stream timeout')
+        recreateLogSession = true
+      else
+        ux.error(message, {exit: 1})
+    }
+  } while (recreateLogSession)
 }
 
 export default logDisplayer
