@@ -117,7 +117,7 @@ describe('heroku certs:auto:enable', function () {
 
     beforeEach(function () {
       clock = lolex.install()
-      clock.setTimeout = function (fn, _timeout) {
+      clock.setTimeout = function (fn) {
         fn()
         return 1
       }
