@@ -10,10 +10,21 @@ describe('pipelines:diff', function () {
     name: 'example-pipeline',
   }
 
+  const firPipeline = {
+    id: '123-pipeline-fir',
+    name: 'example-pipeline',
+  }
+
   const targetApp = {
     id: '123-target-app-456',
     name: 'example-staging',
     pipeline,
+  }
+
+  const targetFirApp = {
+    id: '123-target-app-fir',
+    name: 'example-staging',
+    pipeline: firPipeline,
   }
 
   const targetCoupling = {
@@ -21,6 +32,15 @@ describe('pipelines:diff', function () {
     id: '123-target-app-456',
     pipeline,
     stage: 'staging',
+    generation: 'cedar',
+  }
+
+  const targetFirCoupling = {
+    app: targetFirApp,
+    id: '123-target-app-fir',
+    pipeline: firPipeline,
+    stage: 'staging',
+    generation: 'fir',
   }
 
   const targetGithubApp = {
@@ -33,11 +53,26 @@ describe('pipelines:diff', function () {
     pipeline,
   }
 
+  const downstreamFirApp1 = {
+    id: '123-downstream-app-1-fir',
+    name: 'example-production-eu',
+    pipeline: firPipeline,
+  }
+
   const downstreamCoupling1 = {
     app: downstreamApp1,
     id: '123-target-app-456',
     pipeline,
     stage: 'production',
+    generation: 'cedar',
+  }
+
+  const downstreamFirCoupling1 = {
+    app: downstreamFirApp1,
+    id: '123-target-app-fir',
+    pipeline: firPipeline,
+    stage: 'production',
+    generation: 'fir',
   }
 
   const downstreamApp1Github = {
@@ -50,11 +85,26 @@ describe('pipelines:diff', function () {
     pipeline,
   }
 
+  const downstreamFirApp2 = {
+    id: '123-downstream-app-2-fir',
+    name: 'example-production-us',
+    pipeline: firPipeline,
+  }
+
   const downstreamCoupling2 = {
     app: downstreamApp2,
     id: '123-target-app-456',
     pipeline,
     stage: 'production',
+    generation: 'cedar',
+  }
+
+  const downstreamFirCoupling2 = {
+    app: downstreamFirApp2,
+    id: '123-target-app-fir',
+    pipeline: firPipeline,
+    stage: 'production',
+    generation: 'fir',
   }
 
   const downstreamApp2Github = {
@@ -72,6 +122,17 @@ describe('pipelines:diff', function () {
       })
   }
 
+  function mockFirPipelineCoupling(testInstance: typeof test) {
+    return testInstance
+      .stderr()
+      .stdout()
+      .nock(apiUrl, api => {
+        api
+          .get(`/apps/${targetFirApp.name}/pipeline-couplings`)
+          .reply(200, targetFirCoupling)
+      })
+  }
+
   function mockApps(testInstance: typeof test) {
     return testInstance
       .stderr()
@@ -84,6 +145,21 @@ describe('pipelines:diff', function () {
         api
           .post('/filters/apps')
           .reply(200, [targetApp, downstreamApp1, downstreamApp2])
+      })
+  }
+
+  function mockFirApps(testInstance: typeof test) {
+    return testInstance
+      .stderr()
+      .stdout()
+      .nock(apiUrl, api => {
+        api
+          .get(`/pipelines/${firPipeline.id}/pipeline-couplings`)
+          .reply(200, [targetFirCoupling, downstreamFirCoupling1, downstreamFirCoupling2])
+
+        api
+          .post('/filters/apps')
+          .reply(200, [targetFirApp, downstreamFirApp1, downstreamFirApp2])
       })
   }
 
@@ -175,7 +251,7 @@ describe('pipelines:diff', function () {
       .it('should return an error if the target app has no release')
   })
 
-  describe('for valid apps with a pipeline', function () {
+  describe('for valid Cedar apps with a pipeline', function () {
     const targetSlugId = 'target-slug-id'
     const downstreamSlugId = 'downstream-slug-id'
 
@@ -240,6 +316,76 @@ describe('pipelines:diff', function () {
           .reply(404)
       })
       .command(['pipelines:diff', `--app=${targetApp.name}`])
+      .it('should handle non-200 responses from GitHub', ctx => {
+        expect(ctx.stdout).to.contain('unable to perform a diff')
+      })
+  })
+
+  describe('for valid Fir apps with a pipeline', function () {
+    const targetOciImageId = 'oci-image-id'
+    const downstreamOciImageId = 'downstream-oci-image-id'
+
+    const mockedTest = (testInstance: typeof test) => {
+      const t = mockFirPipelineCoupling(mockFirApps(testInstance))
+
+      return t
+        .stderr()
+        .stdout()
+        .nock(kolkrabbiApi, api => {
+          api
+            .get(`/apps/${targetFirApp.id}/github`)
+            .reply(200, targetGithubApp)
+            .get(`/apps/${downstreamFirApp1.id}/github`)
+            .reply(200, downstreamApp1Github)
+            .get(`/apps/${downstreamFirApp2.id}/github`)
+            .reply(200, downstreamApp2Github)
+            .get('/account/github/token')
+            .reply(200, {github: {token: 'github-token'}})
+        })
+        .nock(apiUrl, api => {
+          api
+            .get(`/apps/${targetFirApp.id}/releases`)
+            .reply(200, [{oci_image: {id: targetOciImageId}, status: 'succeeded'}])
+            .get(`/apps/${downstreamFirApp1.id}/releases`)
+            .reply(200, [
+              {status: 'failed'},
+              {oci_image: {id: downstreamOciImageId}, status: 'succeeded'},
+            ])
+        })
+    }
+
+    mockedTest(test)
+      .stdout()
+      .nock(apiUrl, api => {
+        api
+          .get(`/apps/${targetFirApp.id}/oci-images/${targetOciImageId}`)
+          .reply(200, [{commit: 'COMMIT-HASH'}])
+          .get(`/apps/${downstreamFirApp1.id}/oci-images/${downstreamOciImageId}`)
+          .reply(200, [{commit: 'COMMIT-HASH'}])
+      })
+      .command(['pipelines:diff', `--app=${targetFirApp.name}`])
+      .it('should not compare apps if update to date NOR if repo differs', ctx => {
+        expect(ctx.stdout).to.contain(`⬢ ${targetApp.name} is up to date with ⬢ ${downstreamApp1.name}`)
+        expect(ctx.stdout).to.contain(`⬢ ${targetApp.name} was not compared to ⬢ ${downstreamApp2.name}`)
+      })
+
+    const hashes = ['hash-1', 'hash-2']
+
+    mockedTest(test)
+      .stdout()
+      .nock(apiUrl, api => {
+        api
+          .get(`/apps/${targetFirApp.id}/oci-images/${targetOciImageId}`)
+          .reply(200, [{commit: hashes[0]}])
+          .get(`/apps/${downstreamFirApp1.id}/oci-images/${downstreamOciImageId}`)
+          .reply(200, [{commit: hashes[1]}])
+      })
+      .nock(githubApi, api => {
+        api
+          .get(`/repos/${targetGithubApp.repo}/compare/${hashes[1]}...${hashes[0]}`)
+          .reply(404)
+      })
+      .command(['pipelines:diff', `--app=${targetFirApp.name}`])
       .it('should handle non-200 responses from GitHub', ctx => {
         expect(ctx.stdout).to.contain('unable to perform a diff')
       })
