@@ -1,16 +1,73 @@
-import {expect, test} from '@oclif/test'
+import {stderr} from 'stdout-stderr'
+import Cmd from '../../../../src/commands/ps/restart'
+import runCommand from '../../../helpers/runCommand'
+import * as nock from 'nock'
+import heredoc from 'tsheredoc'
+import expectOutput from '../../../helpers/utils/expectOutput'
+import {expect} from 'chai'
+import stripAnsi = require('strip-ansi')
 
 describe('ps:restart', function () {
-  test
-    .stdout()
-    .stderr()
-    .nock('https://api.heroku.com:443', api => api
+  it('restarts all dynos', async function () {
+    nock('https://api.heroku.com')
       .delete('/apps/myapp/dynos')
-      .reply(200),
-    )
-    .command(['ps:restart', '-a', 'myapp'])
-    .it('restarts all dynos', ({stdout, stderr}) => {
-      expect(stdout).to.be.empty
-      expect(stderr).to.contains('Restarting dynos on ⬢ myapp... done\n')
-    })
+      .reply(202)
+
+    await runCommand(Cmd, [
+      '--app',
+      'myapp',
+    ])
+    expectOutput(stderr.output, heredoc(`
+      Restarting all dynos on ⬢ myapp...
+      Restarting all dynos on ⬢ myapp... done
+    `))
+  })
+
+  it('restarts web dynos', async function () {
+    nock('https://api.heroku.com')
+      .delete('/apps/myapp/formations/web')
+      .reply(202)
+
+    await runCommand(Cmd, [
+      '--app',
+      'myapp',
+      '--process-type',
+      'web',
+    ])
+    expectOutput(stderr.output, heredoc(`
+      Restarting all web dynos on ⬢ myapp...
+      Restarting all web dynos on ⬢ myapp... done
+    `))
+  })
+
+  it('restarts a specific dyno', async function () {
+    nock('https://api.heroku.com')
+      .delete('/apps/myapp/dynos/web.1')
+      .reply(202)
+
+    await runCommand(Cmd, [
+      '--app',
+      'myapp',
+      '--dyno-name',
+      'web.1',
+    ])
+    expectOutput(stderr.output, heredoc(`
+      Restarting dyno web.1 on ⬢ myapp...
+      Restarting dyno web.1 on ⬢ myapp... done
+    `))
+  })
+
+  it('emits a warning when passing dyno as an arg', async function () {
+    nock('https://api.heroku.com')
+      .delete('/apps/myapp/dynos/web.1')
+      .reply(202)
+
+    await runCommand(Cmd, [
+      '--app',
+      'myapp',
+      'web.1',
+    ])
+    expect(stripAnsi(stderr.output)).to.include('DYNO is a deprecated argument.')
+    expect(stderr.output).to.include('Restarting dyno web.1 on ⬢ myapp... done')
+  })
 })
