@@ -39,7 +39,7 @@ describe('ps:type', function () {
     ])
 
     expectOutput(stdout.output, heredoc`
-      === Dyno Types
+      === Process Types
 
        Type Size              Qty Cost/hour Max cost/month 
        ──── ───────────────── ─── ───────── ────────────── 
@@ -52,6 +52,7 @@ describe('ps:type', function () {
        web  Performance-L-RAM 1   ~$0.694   $500           
        web  Performance-XL    1   ~$1.042   $750           
        web  Performance-2XL   1   ~$2.083   $1500          
+
       === Dyno Totals
        Type              Total 
        ───────────────── ───── 
@@ -89,11 +90,12 @@ describe('ps:type', function () {
     api.done()
 
     expect(stdout.output).to.eq(heredoc`
-      === Dyno Types
+      === Process Types
 
        Type Size              Qty Cost/hour Max cost/month 
        ──── ───────────────── ─── ───────── ────────────── 
        web  Performance-L-RAM 1   ~$0.694   $500           
+
       === Dyno Totals
 
        Type              Total 
@@ -123,11 +125,12 @@ describe('ps:type', function () {
       'basic',
     ])
 
-    expectOutput(stdout.output, `=== Dyno Types
+    expectOutput(stdout.output, `=== Process Types
  Type   Size  Qty Cost/hour Max cost/month
  ────── ───── ─── ───────── ──────────────
  web    Basic 1   ~$0.010   $7
  worker Basic 2   ~$0.019   $14
+
 === Dyno Totals
  Type  Total
  ───── ─────
@@ -155,11 +158,12 @@ describe('ps:type', function () {
       'worker=standard-2x',
     ])
 
-    expectOutput(stdout.output, `=== Dyno Types
+    expectOutput(stdout.output, `=== Process Types
  Type   Size        Qty Cost/hour Max cost/month
  ────── ─────────── ─── ───────── ──────────────
  web    Standard-1X 1   ~$0.035   $25
  worker Standard-2X 2   ~$0.139   $100
+
 === Dyno Totals
  Type        Total
  ─────────── ─────
@@ -182,11 +186,12 @@ describe('ps:type', function () {
       'myapp',
     ])
 
-    expectOutput(stdout.output, `=== Dyno Types
+    expectOutput(stdout.output, `=== Process Types
  Type Size     Qty Cost/hour Max cost/month
  ──── ──────── ─── ───────── ──────────────
  web  Shield-M 0
  web  Shield-L 0
+
 === Dyno Totals
  Type     Total
  ──────── ─────
